+++ dump/DumpTest/DumpTest.vcxproj.filters
@@ -24,7 +24,7 @@
     <ClCompile Include="TestCA65Formatter.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\ca65Formatter.cpp">
+    <ClCompile Include="..\formatter\CA65Formatter.cpp">
       <Filter>Source Files\AppSrc</Filter>
     </ClCompile>
   </ItemGroup>
